requirements.txt
# dependencies
psycopg2-binary
fastapi
uvicorn[standard]
httpx
python-dotenv
jinja2
python-multipart
aiofiles
itsdangerous
auth0-server-python>=1.0.0b1,<2
